% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_dashboard}
\alias{launch_dashboard}
\title{Launch the Penguin-Dash Shiny App}
\usage{
launch_dashboard()
}
\value{
No return value; opens the app in a browser.
}
\description{
Launches an interactive Shiny dashboard for exploring the Palmer Penguins dataset.
}
\examples{
\dontrun{
launch_dashboard()
}
}
