% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins_clean}
\alias{penguins_clean}
\title{Cleaned Palmer Penguins Dataset}
\format{
A data frame with 333 rows and 8 variables:
\describe{
\item{species}{Penguin species (Adelie, Chinstrap, or Gentoo)}
\item{island}{Island in Palmer Archipelago, Antarctica}
\item{bill_length_mm}{Bill length in millimeters}
\item{bill_depth_mm}{Bill depth in millimeters}
\item{flipper_length_mm}{Flipper length in millimeters}
\item{body_mass_g}{Body mass in grams}
\item{sex}{Penguin sex (female or male)}
\item{year}{Year of observation}
}
}
\source{
\url{https://allisonhorst.github.io/palmerpenguins/}
}
\usage{
penguins_clean
}
\description{
A subset of the original palmerpenguins data with no missing values.
}
\keyword{datasets}
